+++ packages/hive-keychain-ceramic/CHANGELOG.md
@@ -1,6 +1,13 @@
 # @spknetwork/hive-keychain-ceramic
 
+## 0.0.5
+
+### Patch Changes
+
+- Add IDX data utils package and deprecate IDX profile functions in hive-keychain-ceramic
+
 ## 0.0.2
+
 ### Patch Changes
 
 - f857bf1: First Release